refactor: limit bridge source to images/bridge
Use the chat client code already in images/bridge instead of pkg/chat.
Since bridge was the only consumer, we shouldn't have had two copies of
this library at all!

Also switch the go module name for bridge to be based on this repository
