+++ images/bridge/pkg/chat/internal/error_accumulator_test.go
@@ -5,8 +5,8 @@ import (
 	"errors"
 	"testing"
 
-	utils "github.com/ajbouh/bridge/pkg/chat/internal"
-	"github.com/ajbouh/bridge/pkg/chat/internal/test"
+	utils "github.com/ajbouh/substrate/images/bridge/pkg/chat/internal"
+	"github.com/ajbouh/substrate/images/bridge/pkg/chat/internal/test"
 )
 
 func TestErrorAccumulatorBytes(t *testing.T) {
