+++ images/bridge/pkg/chat/internal/test/helpers.go
@@ -1,7 +1,7 @@
 package test
 
 import (
-	"github.com/ajbouh/bridge/pkg/chat/internal/test/checks"
+	"github.com/ajbouh/substrate/images/bridge/pkg/chat/internal/test/checks"
 
 	"net/http"
 	"os"
