+++ images/bridge/pkg/webrtcpeer/rtc_connection.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/ajbouh/bridge/pkg/router"
+	"github.com/ajbouh/substrate/images/bridge/pkg/router"
 	"github.com/pion/rtp"
 	"github.com/pion/webrtc/v3"
 	"github.com/pion/webrtc/v3/pkg/media"
