--- pkg/chat/README.md
@@ -1,3 +1,0 @@
-# A simplified go client for models that emulate the OpenAI API
-
-Forked from github.com/sashabaranov/go-openai
