--- pkg/chat/chat.go
@@ -1,153 +1,0 @@
-package chat
-
-import (
-	"context"
-	"errors"
-	"net/http"
-)
-
-// Chat message role defined by the OpenAI API.
-const (
-	ChatMessageRoleSystem    = "system"
-	ChatMessageRoleUser      = "user"
-	ChatMessageRoleAssistant = "assistant"
-	ChatMessageRoleFunction  = "function"
-)
-
-const chatCompletionsSuffix = "/chat/completions"
-
-var (
-	ErrChatCompletionStreamNotSupported = errors.New("streaming is not supported with this method, please use CreateChatCompletionStream") //nolint:lll
-)
-
-type PromptAnnotation struct {
-	PromptIndex int `json:"prompt_index,omitempty"`
-}
-
-type ChatCompletionMessage struct {
-	Role    string `json:"role"`
-	Content string `json:"content"`
-
-	// This property isn't in the official documentation, but it's in
-	// the documentation for the official library for python:
-	// - https://github.com/openai/openai-python/blob/main/chatml.md
-	// - https://github.com/openai/openai-cookbook/blob/main/examples/How_to_count_tokens_with_tiktoken.ipynb
-	Name string `json:"name,omitempty"`
-
-	FunctionCall *FunctionCall `json:"function_call,omitempty"`
-}
-
-func (m *ChatCompletionMessage) TokenLength() int {
-	strlen := len(m.Name) + 2 + len(m.Content)
-	if m.FunctionCall != nil {
-		// increase by extra JSON overhead
-		strlen += 9 + len(m.FunctionCall.Name) + len(m.FunctionCall.Arguments)
-	}
-	return strlen / 3 // HACK this is an estimate!
-}
-
-type FunctionCall struct {
-	Name string `json:"name,omitempty"`
-	// call function with arguments in JSON format
-	Arguments string `json:"arguments,omitempty"`
-}
-
-// ChatCompletionRequest represents a request structure for chat completion API.
-type ChatCompletionRequest struct {
-	Model            string                  `json:"model"`
-	Messages         []ChatCompletionMessage `json:"messages"`
-	MaxTokens        int                     `json:"max_tokens,omitempty"`
-	Temperature      float32                 `json:"temperature,omitempty"`
-	TopP             float32                 `json:"top_p,omitempty"`
-	N                int                     `json:"n,omitempty"`
-	Stream           bool                    `json:"stream,omitempty"`
-	Stop             []string                `json:"stop,omitempty"`
-	PresencePenalty  float32                 `json:"presence_penalty,omitempty"`
-	FrequencyPenalty float32                 `json:"frequency_penalty,omitempty"`
-	// LogitBias is must be a token id string (specified by their token ID in the tokenizer), not a word string.
-	// incorrect: `"logit_bias":{"You": 6}`, correct: `"logit_bias":{"1639": 6}`
-	// refs: https://platform.openai.com/docs/api-reference/chat/create#chat/create-logit_bias
-	LogitBias    map[string]int       `json:"logit_bias,omitempty"`
-	User         string               `json:"user,omitempty"`
-	Functions    []FunctionDefinition `json:"functions,omitempty"`
-	FunctionCall any                  `json:"function_call,omitempty"`
-}
-
-func (r *ChatCompletionRequest) InsertMessagesAt(i int, more ...ChatCompletionMessage) {
-	r.Messages = append(r.Messages[:i], append(more, r.Messages[i:]...)...)
-}
-
-type FunctionDefinition struct {
-	Name        string `json:"name"`
-	Description string `json:"description,omitempty"`
-	// Parameters is an object describing the function.
-	// You can pass json.RawMessage to describe the schema,
-	// or you can pass in a struct which serializes to the proper JSON schema.
-	// The jsonschema package is provided for convenience, but you should
-	// consider another specialized library if you require more complex schemas.
-	Parameters any `json:"parameters"`
-}
-
-// Deprecated: use FunctionDefinition instead.
-type FunctionDefine = FunctionDefinition
-
-type FinishReason string
-
-const (
-	FinishReasonStop          FinishReason = "stop"
-	FinishReasonLength        FinishReason = "length"
-	FinishReasonFunctionCall  FinishReason = "function_call"
-	FinishReasonContentFilter FinishReason = "content_filter"
-	FinishReasonNull          FinishReason = "null"
-)
-
-func (r FinishReason) MarshalJSON() ([]byte, error) {
-	if r == FinishReasonNull || r == "" {
-		return []byte("null"), nil
-	}
-	return []byte(`"` + string(r) + `"`), nil // best effort to not break future API changes
-}
-
-type ChatCompletionChoice struct {
-	Index   int                   `json:"index"`
-	Message ChatCompletionMessage `json:"message"`
-	// FinishReason
-	// stop: API returned complete message,
-	// or a message terminated by one of the stop sequences provided via the stop parameter
-	// length: Incomplete model output due to max_tokens parameter or token limit
-	// function_call: The model decided to call a function
-	// content_filter: Omitted content due to a flag from our content filters
-	// null: API response still in progress or incomplete
-	FinishReason FinishReason `json:"finish_reason"`
-}
-
-// ChatCompletionResponse represents a response structure for chat completion API.
-type ChatCompletionResponse struct {
-	ID      string                 `json:"id"`
-	Object  string                 `json:"object"`
-	Created int64                  `json:"created"`
-	Model   string                 `json:"model"`
-	Choices []ChatCompletionChoice `json:"choices"`
-	Usage   Usage                  `json:"usage"`
-}
-
-// CreateChatCompletion — API call to Create a completion for the chat message.
-func (c *Client) CreateChatCompletion(
-	ctx context.Context,
-	request ChatCompletionRequest,
-) (response ChatCompletionResponse, err error) {
-	if request.Stream {
-		err = ErrChatCompletionStreamNotSupported
-		return
-	}
-
-	urlSuffix := chatCompletionsSuffix
-
-	req, err := c.newRequest(ctx, http.MethodPost, c.fullURL(urlSuffix, request.Model), withBody(request))
-	if err != nil {
-		return
-	}
-
-	err = c.sendRequest(req, &response)
-	return
-}
