--- pkg/chat/client_test.go
@@ -1,193 +1,0 @@
-package chat //nolint:testpackage // testing private field
-
-import (
-	"bytes"
-	"context"
-	"errors"
-	"fmt"
-	"io"
-	"net/http"
-	"testing"
-
-	"github.com/ajbouh/substrate/pkg/chat/internal/test"
-)
-
-var errTestRequestBuilderFailed = errors.New("test request builder failed")
-
-type failingRequestBuilder struct{}
-
-func (*failingRequestBuilder) Build(_ context.Context, _, _ string, _ any, _ http.Header) (*http.Request, error) {
-	return nil, errTestRequestBuilderFailed
-}
-
-func TestDecodeResponse(t *testing.T) {
-	stringInput := ""
-
-	testCases := []struct {
-		name     string
-		value    interface{}
-		body     io.Reader
-		hasError bool
-	}{
-		{
-			name:  "nil input",
-			value: nil,
-			body:  bytes.NewReader([]byte("")),
-		},
-		{
-			name:  "string input",
-			value: &stringInput,
-			body:  bytes.NewReader([]byte("test")),
-		},
-		{
-			name:  "map input",
-			value: &map[string]interface{}{},
-			body:  bytes.NewReader([]byte(`{"test": "test"}`)),
-		},
-		{
-			name:     "reader return error",
-			value:    &stringInput,
-			body:     &errorReader{err: errors.New("dummy")},
-			hasError: true,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			err := decodeResponse(tc.body, tc.value)
-			if (err != nil) != tc.hasError {
-				t.Errorf("Unexpected error: %v", err)
-			}
-		})
-	}
-}
-
-type errorReader struct {
-	err error
-}
-
-func (e *errorReader) Read(_ []byte) (n int, err error) {
-	return 0, e.err
-}
-
-func TestHandleErrorResp(t *testing.T) {
-	// var errRes *ErrorResponse
-	var errRes ErrorResponse
-	var reqErr RequestError
-	t.Log(errRes, errRes.Error)
-	if errRes.Error != nil {
-		reqErr.Err = errRes.Error
-	}
-	t.Log(fmt.Errorf("error, %w", &reqErr))
-	t.Log(errRes.Error, "nil pointer check Pass")
-
-	const mockToken = "mock token"
-	client := NewClient(mockToken)
-
-	testCases := []struct {
-		name     string
-		httpCode int
-		body     io.Reader
-		expected string
-	}{
-		{
-			name:     "401 Invalid Authentication",
-			httpCode: http.StatusUnauthorized,
-			body: bytes.NewReader([]byte(
-				`{
-					"error":{
-						"message":"You didn't provide an API key. ....",
-						"type":"invalid_request_error",
-						"param":null,
-						"code":null
-					}
-				}`,
-			)),
-			expected: "error, status code: 401, message: You didn't provide an API key. ....",
-		},
-		{
-			name:     "401 Azure Access Denied",
-			httpCode: http.StatusUnauthorized,
-			body: bytes.NewReader([]byte(
-				`{
-					"error":{
-						"code":"AccessDenied",
-						"message":"Access denied due to Virtual Network/Firewall rules."
-					}
-				}`,
-			)),
-			expected: "error, status code: 401, message: Access denied due to Virtual Network/Firewall rules.",
-		},
-		{
-			name:     "503 Model Overloaded",
-			httpCode: http.StatusServiceUnavailable,
-			body: bytes.NewReader([]byte(`
-				{
-					"error":{
-						"message":"That model...",
-						"type":"server_error",
-						"param":null,
-						"code":null
-					}
-				}`)),
-			expected: "error, status code: 503, message: That model...",
-		},
-		{
-			name:     "503 no message (Unknown response)",
-			httpCode: http.StatusServiceUnavailable,
-			body: bytes.NewReader([]byte(`
-				{
-					"error":{}
-				}`)),
-			expected: "error, status code: 503, message: ",
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			testCase := &http.Response{}
-			testCase.StatusCode = tc.httpCode
-			testCase.Body = io.NopCloser(tc.body)
-			err := client.handleErrorResp(testCase)
-			t.Log(err.Error())
-			if err.Error() != tc.expected {
-				t.Errorf("Unexpected error: %v , expected: %s", err, tc.expected)
-				t.Fail()
-			}
-
-			e := &APIError{}
-			if !errors.As(err, &e) {
-				t.Errorf("(%s) Expected error to be of type APIError", tc.name)
-				t.Fail()
-			}
-		})
-	}
-}
-
-func TestClientReturnsRequestBuilderErrors(t *testing.T) {
-	config := DefaultConfig(test.GetTestToken())
-	client := NewClientWithConfig(config)
-	client.requestBuilder = &failingRequestBuilder{}
-	ctx := context.Background()
-
-	type TestCase struct {
-		Name     string
-		TestFunc func() (any, error)
-	}
-
-	testCases := []TestCase{
-		{"CreateChatCompletion", func() (any, error) {
-			return client.CreateChatCompletion(ctx, ChatCompletionRequest{Model: "GPT3Dot5Turbo"})
-		}},
-		{"CreateChatCompletionStream", func() (any, error) {
-			return client.CreateChatCompletionStream(ctx, ChatCompletionRequest{Model: "GPT3Dot5Turbo"})
-		}},
-	}
-
-	for _, testCase := range testCases {
-		_, err := testCase.TestFunc()
-		if !errors.Is(err, errTestRequestBuilderFailed) {
-			t.Fatalf("%s did not return error when request builder failed: %v", testCase.Name, err)
-		}
-	}
-}
