--- pkg/chat/common.go
@@ -1,10 +1,0 @@
-package chat
-
-// common.go defines common types used throughout the OpenAI API.
-
-// Usage Represents the total token usage per request to OpenAI.
-type Usage struct {
-	PromptTokens     int `json:"prompt_tokens"`
-	CompletionTokens int `json:"completion_tokens"`
-	TotalTokens      int `json:"total_tokens"`
-}
