--- pkg/chat/internal/error_accumulator_test.go
@@ -1,41 +1,0 @@
-package internal_test
-
-import (
-	"bytes"
-	"errors"
-	"testing"
-
-	utils "github.com/ajbouh/substrate/pkg/chat/internal"
-	"github.com/ajbouh/substrate/pkg/chat/internal/test"
-)
-
-func TestErrorAccumulatorBytes(t *testing.T) {
-	accumulator := &utils.DefaultErrorAccumulator{
-		Buffer: &bytes.Buffer{},
-	}
-
-	errBytes := accumulator.Bytes()
-	if len(errBytes) != 0 {
-		t.Fatalf("Did not return nil with empty bytes: %s", string(errBytes))
-	}
-
-	err := accumulator.Write([]byte("{}"))
-	if err != nil {
-		t.Fatalf("%+v", err)
-	}
-
-	errBytes = accumulator.Bytes()
-	if len(errBytes) == 0 {
-		t.Fatalf("Did not return error bytes when has error: %s", string(errBytes))
-	}
-}
-
-func TestErrorByteWriteErrors(t *testing.T) {
-	accumulator := &utils.DefaultErrorAccumulator{
-		Buffer: &test.FailingErrorBuffer{},
-	}
-	err := accumulator.Write([]byte("{"))
-	if !errors.Is(err, test.ErrTestErrorAccumulatorWriteFailed) {
-		t.Fatalf("Did not return error when write failed: %v", err)
-	}
-}
