--- pkg/chat/internal/marshaller.go
@@ -1,15 +1,0 @@
-package internal
-
-import (
-	"encoding/json"
-)
-
-type Marshaller interface {
-	Marshal(value any) ([]byte, error)
-}
-
-type JSONMarshaller struct{}
-
-func (jm *JSONMarshaller) Marshal(value any) ([]byte, error) {
-	return json.Marshal(value)
-}
