--- pkg/chat/internal/unmarshaler.go
@@ -1,15 +1,0 @@
-package internal
-
-import (
-	"encoding/json"
-)
-
-type Unmarshaler interface {
-	Unmarshal(data []byte, v any) error
-}
-
-type JSONUnmarshaler struct{}
-
-func (jm *JSONUnmarshaler) Unmarshal(data []byte, v any) error {
-	return json.Unmarshal(data, v)
-}
